Implemented loading #2 and displaying #3 of heightmap

+++ src/framework/model.cpp
@@ -46,11 +46,11 @@ namespace framework
                                    attrib.vertices[3 * index.vertex_index + 1],
                                    attrib.vertices[3 * index.vertex_index + 2] };
 
-                vertex.nor = { attrib.normals[static_cast<__int64>(3) * index.normal_index + 0],
+                vertex.nor = { attrib.normals[3 * index.normal_index + 0],
                                   attrib.normals[3 * index.normal_index + 1],
                                   attrib.normals[3 * index.normal_index + 2] };
 
-                vertex.tex = { attrib.texcoords[static_cast<__int64>(2) * index.texcoord_index + 0],
+                vertex.tex = { attrib.texcoords[2 * index.texcoord_index + 0],
                                    attrib.texcoords[2 * index.texcoord_index + 1] };
 
                 if (vertex_cache.find(vertex) == vertex_cache.end()) // Checking if vertex isn't in cache
